requirements.txt
# RSS AI Digest - Dependencies

# RSS/Atom parsing
feedparser>=6.0.0

# Article content extraction
trafilatura>=1.6.0

# AI API clients
anthropic>=0.25.0
openai>=1.0.0

# Utilities
pyyaml>=6.0.0
requests>=2.31.0
python-slugify>=8.0.0

# Static site generation
jinja2>=3.1.0